Refactor and add styling
